Better error handling in NewSalesOrderService

+++ src/CPE.Sales/ViewModel/ParserSettingsTestViewModel.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using CPE.Domain.Services;
 using CPE.Sales.Services;
 using GalaSoft.MvvmLight;
 
@@ -12,6 +13,9 @@ namespace CPE.Sales.ViewModel
     public sealed class ParserSettingsTestViewModel : ViewModelBase
     {
         private NewSalesOrdersService _salesOrderService;
+        private ICustomerService _customers;
+        private ITricornService _tricorn;
+
         private string _statusMessage = "Please ensure settings are correct then press the 'Scan folder now' button.";
 
         public string StatusMessage
@@ -24,11 +28,44 @@ namespace CPE.Sales.ViewModel
             }
         }
 
-        public ParserSettingsTestViewModel(NewSalesOrdersService salesOrderService)
+        public ParserSettingsTestViewModel(NewSalesOrdersService salesOrderService, ICustomerService customers, ITricornService tricorn)
         {
             _salesOrderService = salesOrderService;
+            _tricorn = tricorn;
+            _customers = customers;
         }
 
+        public async Task CheckCpeCentralConnection()
+        {
+            StatusMessage = "Checking....";
+
+            try
+            {
+                var c = await _customers.GetAllAsync();
+
+                StatusMessage = "Connected to CPECentral successfully!";
+            }
+            catch (Exception ex)
+            {
+                StatusMessage = ex.Message;
+            } 
+        }
+
+        public async Task CheckTricornConnection()
+        {
+            StatusMessage = "Checking....";
+
+            try
+            {
+                var c = await _tricorn.GetNameByDrawingNumberAsync("H17070A");
+
+                StatusMessage = "Connected to Tricorn successfully!";
+            }
+            catch (Exception ex)
+            {
+                StatusMessage = ex.Message;
+            }
+        }
         public async Task PerformTestAsync(string orderDirectory, string outputDirectory)
         {
             StatusMessage = "Scanning sales order folder....";
@@ -49,11 +86,13 @@ namespace CPE.Sales.ViewModel
                 {
                     var fileName = Path.Combine(outputDirectory, Path.GetFileName(so.MailItem.Attachments[0]));
 
-                    File.Copy(so.MailItem.Attachments[0], fileName);
+                    if (!File.Exists(fileName))
+                    {
+                        File.Copy(so.MailItem.Attachments[0], fileName);
+                    }
                 }
 
-                StatusMessage = string.Format("Failed to parse {0} orders. Orders copied to selected directory.",
-                    failedOrders.Count());
+                StatusMessage = $"Failed to parse {failedOrders.Count()} orders. Orders copied to selected directory.";
             }
             else
             {
